svg: filters and pages
Netlistsvg can now be called multiple times for subschematics. This allows
us to draw the schematics in parts.

One of the ways to limit the schematic part is through a netlist name filter.
This way things like i2c maps, power trees, or sections of a chip
(ex: MMC interface) can be drawn separately.

The other way is matching the order of the source code, but with a limit
of pins per "page". This way parts in the generated svg schematics will
approximatelly be in the same order as the code. This mode was added since
netlistsvg kinda randomly places subcircuits, probably based on
best physical fit, hopefully this is a good way to hint it where to put stuff.

Signed-off-by: Alexandru M Stan <amstan@chromium.org>