test: servo_micro's allegro output compared with the original netlist
As I might have said earlier, servo_micro originally implemented in Allegro (and that's still the source of truth).
What better way to do a sanity check than to compare the output of PCBDL to a netlist from the real design.

test/integration/servo_original.rpt is the netlist exported from the original Allegro files:
* internal git hash 9447552346992fbb1690e8d185410ae2ba87a44e
* 652-00166-05_20190308_servo_micro_SCH_0.pdf should match it

Added test/integration/netlist_normalize.py as a tool to make sure the irrelevant details from the netlist
(ordering, resistor pins, part_numbers) won't pollute the diff. Ideally the normalized outputs can be
compared and when they're identical we have a pretty good confidence that:
    * servo_micro was re-implemented properly in PCBDL
    * PCBDL is still sane. As I fiddle with more core language features this would be a good sanity
      check that behavior doesn't differ. The netlist output should be stable, even while the language
      is updated in the background.

test/integration/netlist.py is the main integration test this.
    * loads servo_micro
        * generates the allegro netlist for it
    * normalizes both our output and servo_original.rpt
    * asserts if there's differences

BUG=Issue #22
TEST="make test", See if fail because our servo_micro implementation is still not correct