base: Swap the position of PinFragment numbers and names
Pin numbers are way more likely to be the same width, so when entering a large pin count,
especially when some pins have a lot of names, the numbers can get lost at the end of the line.

Swap the position so the numbers are first.

examples/servo_micro has also been updated for this changes.

Documentation for PinFragment to follow.

TEST=test/integration/netlist.py